INS-2588: moce gatewayer code from servicenetwork to separate file
